Split code into several modules
