+++ src/extra_writer/__init__.py
@@ -35,7 +35,14 @@ You should have received a copy of the 3-Clause BSD License along with this
 program. If not, see <https://opensource.org/licenses/BSD-3-Clause>
 """
 
-from .version import __version__
-from .writer import FileWriter, DS, trs_
+from . import version
+from .writer import FileWriter
+from .dataset_descriptor import Dataset, MultiTrainData
+
+__version__ = version.version
+
+DS = Dataset
+trs_ = MultiTrainData
+
 
 __all__ = ["FileWriter", "DS", "trs_"]
