+++ src/extra_writer/version.py
@@ -1,1 +1,1 @@
-__version__ = '0.0.1'
+version = '0.0.1'
