+++ src/extra_writer/writer.py
@@ -1,364 +1,7 @@
-# compatibility to future features
-from . import future
-
 import h5py
 import numpy as np
 
-from itertools import accumulate
-
-
-class DatasetWriterBase:
-    """Abstract class for writers wrapping h5py"""
-    def __init__(self, ds, file_ds, chunks):
-        self.ds = ds
-        self.file_ds = file_ds
-        self.chunks = chunks
-        self.pos = 0
-
-    def flush(self):
-        pass
-
-    def write(self, data, nrec, start=None):
-        raise NotImplementedError
-
-
-class DatasetDirectWriter(DatasetWriterBase):
-    """Class writes data directly to the file"""
-    def write(self, data, nrec, start=None):
-        """Writes data to disk"""
-        end = self.pos + nrec
-        if start is None:
-            start = 0
-        self.file_ds.resize(end, 0)
-        if not np.ndim(data):
-            self.file_ds[self.pos:end] = data
-        else:
-            # self.file_ds[self.pos:end] = data[start:start+nrec]
-            self.file_ds.write_direct(
-                data, np.s_[start:start+nrec], np.s_[self.pos:end])
-
-
-class DatasetBufferedWriter(DatasetWriterBase):
-    """Class implements buffered writing"""
-    def __init__(self, ds, file_ds, chunks):
-        super().__init__(ds, file_ds, chunks)
-        self._data = np.empty(chunks, dtype=ds.dtype)
-        self.size = chunks[0]
-        self.nbuf = 0
-
-    def flush(self):
-        """Writes buffer to disk"""
-        if self.nbuf:
-            end = self.pos + self.nbuf
-            self.file_ds.resize(end, 0)
-            self.file_ds.write_direct(
-                self._data, np.s_[:self.nbuf], np.s_[self.pos:end])
-            self.pos = end
-            self.nbuf = 0
-
-    def write_one(self, value):
-        """Buffers single record"""
-        self._data[self.nbuf] = value
-        self.nbuf += 1
-        if self.nbuf >= self.size:
-            self.flush()
-
-    def write_many(self, arr, nrec, start=None):
-        """Buffers multiple records"""
-        if start is None:
-            start = 0
-        buf_nrest = self.size - self.nbuf
-        data_nrest = nrec - buf_nrest
-        if data_nrest < 0:
-            # copy
-            end = self.nbuf + nrec
-            self._data[self.nbuf:end] = arr[start:start+nrec]
-            self.nbuf = end
-        elif self.nbuf and data_nrest < self.size:
-            # copy, flush, copy
-            split = start + buf_nrest
-            self._data[self.nbuf:] = arr[start:split]
-
-            end = self.pos + self.size
-            self.file_ds.resize(end, 0)
-            self.file_ds.write_direct(
-                self._data, np.s_[:], np.s_[self.pos:end])
-            self.pos = end
-
-            self._data[:data_nrest] = arr[split:start+nrec]
-            self.nbuf = data_nrest
-        else:
-            # flush, write, copy
-            nrest = nrec % self.size
-            nwrite = nrec - nrest
-
-            split = self.pos + self.nbuf
-            end = split + nwrite
-            self.file_ds.resize(end, 0)
-            if self.nbuf:
-                self.file_ds.write_direct(
-                    self._data, np.s_[:self.nbuf], np.s_[self.pos:split])
-            self.file_ds.write_direct(arr, np.s_[start:start+nwrite],
-                                      np.s_[split:end])
-
-            self._data[:nrest] = arr[start+nwrite:start+nrec]
-            self.pos = end
-            self.nbuf = nrest
-
-    def write(self, data, nrec, start=None):
-        """Buffer or writes data"""
-        if nrec == 1:
-            self.write_one(data[start] if start is not None else data)
-        else:
-            if not isinstance(data, np.ndarray):
-                arr = np.array(data, dtype=self.ds.dtype)
-            else:
-                arr = data
-
-            # arr = np.broadcast_to(data, (nrec,) + self.ds.entry_shape)
-            self.write_many(arr, nrec, start=start)
-
-
-class DataQueue:
-    def __init__(self):
-        self.data = []
-        self.size = []
-        self.nwritten = 0
-        self.nready = 0
-
-    def __bool__(self):
-        return bool(self.data)
-
-    def append(self, count, data):
-        """Appends data into queue for future writing"""
-        ntrain = len(count)
-        self.data.append((count, data))
-        self.size.append((0, 0, ntrain))
-        self.nready += ntrain
-
-    def reset(self):
-        """Reset counters for new file"""
-        self.nready -= self.nwritten
-        self.nwritten = 0
-
-    def get_max_below(self, end):
-        """Finds the maximum number of trains that data items in the queue
-        fill without splitting and below the given number"""
-        ntrain = self.nwritten
-        for _, _, n in self.size:
-            if ntrain + n > end:
-                return ntrain
-            ntrain += n
-        return ntrain
-
-    def write(self, writer, end):
-        """Writes data items from the queue"""
-        nrest = end - self.nwritten
-        while nrest > 0:
-            offset, train0, ntrain = self.size[0]
-            if ntrain == 1 and offset == 0:
-                count, data = self.data.pop(0)
-                writer.write(data, count[0])
-                nrest -= 1
-                self.size.pop(0)
-            elif ntrain <= nrest:
-                count, data = self.data.pop(0)
-                trainN = train0 + ntrain
-                nrec = np.sum(count[train0:trainN])
-                writer.write(data, nrec, offset)
-
-                nrest -= ntrain
-                self.size.pop(0)
-            else:
-                count, data = self.data[0]
-                trainN = train0 + nrest
-                nrec = np.sum(count[train0:trainN])
-                writer.write(data, nrec, offset)
-
-                self.size[0] = (offset + nrec, trainN, ntrain - nrest)
-                nrest = 0
-
-        self.nwritten = end
-
-
-# Attention! Do not instanciate `Source` in the metaclass `FileWriterMeta`
-class Source:
-    """Creates data source group and its indexes"""
-
-    SECTION = ('CONTROL', 'INSTRUMENT')
-
-    def __init__(self, writer, name, stype=None):
-        self.writer = writer
-        self.name = name
-        if stype is None:
-            self.stype = int(':' in name)
-        else:
-            self.stype = stype
-
-        self.section = self.SECTION[self.stype]
-        if writer._meta.break_into_sequence:
-            self.max_trains = writer._meta.max_train_per_file
-        else:
-            self.max_trains = None
-
-        self.ndatasets = 0
-        self.nready = 0
-
-        self.datasets = []
-        self.dsno = {}
-        self.file_ds = []
-        self.data = []
-
-        self.count_buf = []
-
-        self.first = []
-        self.count = []
-        self.nrec = 0
-
-        self.block_writing = True
-
-    def add(self, name, ds):
-        """Adds dataset to the source"""
-        self.dsno[name] = len(self.datasets)
-        self.datasets.append(ds)
-        self.data.append(DataQueue())
-        self.file_ds.append(None)
-        self.ndatasets += 1
-
-    def create(self):
-        """Creates all datasets in file"""
-        grp = self.writer._file.create_group(self.section + '/' + self.name)
-        for dsno, ds in enumerate(self.datasets):
-            self.file_ds[dsno] = ds.create(self.writer, grp)
-        self._grp = grp
-        self.block_writing = False
-
-        while self.nready >= self.ndatasets and not self.block_writing:
-            self.write_data()
-
-        return grp
-
-    def create_index(self, index_grp, max_trains):
-        """Create source index in h5-file"""
-        grp = index_grp.create_group(self.name)
-        for key in ('first', 'count'):
-            ds = grp.create_dataset(
-                key, (max_trains,), dtype=np.uint64, chunks=(max_trains,),
-                maxshape=(None,)
-            )
-            ds[:] = 0
-
-    def write_index(self, index_grp, ntrains):
-        """Writes source index in h5-file"""
-        nmissed = ntrains - len(self.count)
-        if nmissed > 0:
-            self.count += [0] * nmissed
-            self.first += [self.nrec] * nmissed
-
-        grp = index_grp[self.name]
-        for dsname in ('first', 'count'):
-            ds = grp[dsname]
-            ds.resize(ntrains, axis=0)
-            val = getattr(self, dsname)
-            ds[:] = val[:ntrains]
-            del val[:ntrains]
-
-        if len(self.count):
-            self.first = list(accumulate(self.count[:-1], initial=0))
-            self.nrec = sum(self.count)
-        else:
-            self.first = []
-            self.nrec = 0
-
-        del self.count_buf[:ntrains]
-
-    def close_datasets(self):
-        """Finalize writing"""
-        for dsno, ds in enumerate(self.datasets):
-            self.file_ds[dsno].flush()
-            self.data[dsno].reset()
-
-    def add_train_data(self, nrec, name, value):
-        """Adds single train data to the source"""
-        if self.stype == 0 and nrec != 1:
-            raise ValueError("maximum one entry per train can be written "
-                             "in control source")
-
-        dsno = self.dsno[name]
-
-        ntrain = self.data[dsno].nready
-        nwritten = len(self.count)
-        if nwritten > ntrain:
-            if self.count_buf[ntrain] != nrec:
-                raise ValueError("count mismatch the number of frames "
-                                 "in source")
-        else:
-            self.count_buf.append(nrec)
-
-        self._put_data(dsno, [nrec], value)
-
-    def add_data(self, count, name, value):
-        """Adds multitrain data to the source"""
-        if self.stype == 0 and np.any(count > 1):
-            raise ValueError("maximum one entry per train can be written "
-                             "in control source")
-
-        dsno = self.dsno[name]
-
-        ntrain = self.data[dsno].nready
-        nwritten = len(self.count)
-        if nwritten > ntrain:
-            nmatch = min(nwritten - ntrain, len(count))
-            if np.any(self.count_buf[ntrain:ntrain+nmatch] != count[:nmatch]):
-                raise ValueError("count mismatch the number of frames "
-                                 "in source")
-
-            self.count_buf += count[nmatch:].tolist()
-        else:
-            self.count_buf += count.tolist()
-
-        self._put_data(dsno, count, value)
-
-    def _put_data(self, dsno, count, value):
-        self.nready += not self.data[dsno]
-        self.data[dsno].append(count, value)
-
-        while self.nready >= self.ndatasets and not self.block_writing:
-            self.write_data()
-
-    def write_data(self):
-        """Write data when the trains completely filled"""
-        train0 = len(self.count)
-        max_ready = min(d.nready for d in self.data)
-
-        if self.max_trains is not None and self.max_trains < max_ready:
-            max_ready = self.max_trains
-            self.block_writing = True
-
-        self.nready = 0
-        trainN = train0
-        for dsno in range(self.ndatasets):
-            if self.block_writing:
-                end = max_ready
-            else:
-                end = self.data[dsno].get_max_below(max_ready)
-
-            self.data[dsno].write(self.file_ds[dsno], end)
-            self.nready += bool(self.data[dsno])
-            trainN = max(trainN, end)
-
-        count = self.count_buf[train0:trainN]
-        first = list(accumulate(count[:-1], initial=self.nrec))
-        self.count += count
-        self.first += first
-        self.nrec += np.sum(count, dtype=int)
-
-    def get_ntrain(self, dsname):
-        dsno = self.dsno[dsname]
-        return self.data[dsno].nready
-
-    def get_min_trains(self):
-        return min(d.nready for d in self.data)
+from .dataset_descriptor import DatasetBase, Dataset, BlockedSetter
 
 
 class Options:
@@ -406,217 +49,6 @@ class Options:
         if meta_attrs != {}:
             raise TypeError("'class Meta' got invalid attribute(s): " +
                             ','.join(meta_attrs))
-
-
-class MultiTrainData:
-    def __init__(self, count, data):
-        self.count = count
-        self.data = data
-
-
-class DataSetterBase:
-    def __set__(self, instance, value):
-        raise NotImplementedError
-
-
-class BlockedSetter(DataSetterBase):
-    def __set__(self, instance, value):
-        raise RuntimeError(
-            "Class attributes interface is disabled. Use option "
-            "'class_attrs_interface=True' to enable it.")
-
-
-class DataSetter(DataSetterBase):
-    """Overrides the setters for attributes which declared as datasets
-    in order to use the assignment operation for adding data in a train
-    """
-    def __init__(self, name):
-        self.name = name
-
-    def __set__(self, instance, value):
-        if isinstance(value, MultiTrainData):
-            instance.add_value(value.count, self.name, value.data)
-        else:
-            instance.add_train_value(self.name, value)
-
-
-# Attention! `Dataset` is the descriptor class and its instances are
-# intended to be used as class members of `FileWriter` children. Changing
-# them leads to changes in the host class itself and in all its instances.
-# Therefore, one can change `self` only in the `__init__` method and
-# in methods that are called from the `FileWriterMeta` metaclass.
-class DatasetBase:
-    """Base dataset descriptor class"""
-    source_class = None
-
-    class Attributes:
-        """Dataset attributes structure"""
-        def __init__(self, **kwargs):
-            for kw, val in kwargs.items():
-                setattr(self, kw, val)
-
-    def __init__(self, source_name, key, entry_shape, dtype,
-                 chunks=None, compression=None):
-        self.entry_shape = entry_shape
-        self.dtype = dtype
-        self.compression = compression
-        self.chunks = chunks
-        if source_name and source_name[0] == '@':
-            self.orig_name = (True, source_name[1:], key)
-        else:
-            self.orig_name = (False, source_name, key)
-
-    def set_name(self, source_name, key):
-        """Sets new name to the dataset"""
-        self.orig_name = (False, source_name, key)
-        self.canonical_name = (source_name, key)
-
-    def resolve_name(self, aliases={}):
-        """Normalizes source name and key"""
-        isalias, source_id, key = self.orig_name
-
-        # resolve reference
-        source_name = aliases[source_id] if isalias else source_id
-        self.canonical_name = (source_name, key)
-
-    def get_dataset_fullname(self, writer):
-        # expected to return (source_name, key, stype)
-        raise NotImplementedError
-
-    def get_entry_attr(self, writer):
-        # expected to return (entry_shape, dtype)
-        return self.entry_shape, self.dtype
-
-    def get_chunks(self, writer):
-        # expected to return chunks
-        return self.chunks
-
-    def init_dataset_attr(self, writer):
-        # expand names
-        source_name, key, stype = self.get_dataset_fullname(writer)
-
-        # expand entry attributes
-        entry_shape, dtype = self.get_entry_attr(writer)
-
-        # auto chunking
-        chunks = self.get_chunks(writer)
-        if chunks is None:
-            chunks = Dataset._chunks_autosize(
-                writer._meta.max_train_per_file, entry_shape,
-                dtype, stype)
-
-        writer._ds_attrs[id(self)] = Dataset.Attributes(
-            source_name=source_name, key=key, stype=stype,
-            entry_shape=entry_shape, dtype=dtype, chunks=chunks,
-            compression=self.compression,
-        )
-
-    def check_value(self, writer, value):
-        """Checks data"""
-        # can we need to checked type cast?
-
-        # shape check
-        entry_shape = self(writer).entry_shape
-        value_shape = np.shape(value)
-        shape = np.broadcast_shapes(value_shape, (1,) + entry_shape)
-        if shape == entry_shape:
-            nrec = 1
-        elif shape[1:] == entry_shape:
-            nrec = shape[0]
-        else:
-            raise ValueError(f"shape mismatch: {value_shape} cannot "
-                             f"be broadcast to {(None, ) + entry_shape}")
-        return nrec
-
-    @staticmethod
-    def _chunks_autosize(max_trains, entry_shape, dtype, stype):
-        """Caclulates chunk size"""
-        MN = (max_trains, 32, 32)
-        SZ = (1 << 14, 1 << 19, 1 << 23)  # 16K, 512K, 8M
-
-        size = np.prod(entry_shape, dtype=int)
-        ndim = len(entry_shape)
-        nbytes = size * np.dtype(dtype).itemsize
-
-        entry_type = int(size != 1) * (1 + int(ndim > 1))
-        chunk = max(SZ[entry_type] // nbytes, MN[entry_type])
-        if stype == 0:
-            chunk = min(chunk, max_trains)
-
-        return (chunk,) + tuple(entry_shape)
-
-    def __call__(self, writer):
-        return writer._ds_attrs[id(self)]
-
-    def create(self, writer, grp):
-        """Creates dataset in h5-file and return writer"""
-        attr = self(writer)
-
-        ds = grp.create_dataset(
-            attr.key, (0,) + attr.entry_shape,
-            dtype=attr.dtype, chunks=attr.chunks, maxshape=(None,)
-            + attr.entry_shape, compression=attr.compression
-        )
-        if writer._meta.buffering:
-            wrt = DatasetBufferedWriter(self, ds, attr.chunks)
-        else:
-            wrt = DatasetDirectWriter(self, ds, attr.chunks)
-
-        return wrt
-
-    def source(self, writer):
-        raise NotImplementedError
-
-
-class Dataset(DatasetBase):
-    """Dataset descriptor"""
-    source_class = Source
-
-    def source(self, writer):
-        attr = self(writer)
-        return self.source_class(writer, attr.source_name, attr.stype)
-
-    def get_attribute_setter(self, name):
-        """Returns suitable attribute setter instance"""
-        return DataSetter(name)
-
-    def get_dataset_fullname(self, writer):
-        # expected to return (source_name, key, stype)
-        source_name, key = self.canonical_name
-        return Dataset._normalize_name(
-            source_name.format(**writer.param), key)
-
-    def get_chunks(self, writer):
-        # expected to return chunks
-        return Dataset._expand_shape(self.chunks, writer)
-
-    def get_entry_attr(self, writer):
-        # expected to return (entry_shape, dtype)
-        return (Dataset._expand_shape(self.entry_shape, writer),
-                self.dtype)
-
-    @staticmethod
-    def _normalize_name(source_name, key):
-        """Transforms canonical name to the internal form"""
-        # can we really distinguish sources by colons?
-        stype = int(':' in source_name)
-        if stype:
-            tk, key = key.split('.', 1)
-            source_name = source_name + '/' + tk
-
-        return source_name, key.replace('.', '/'), stype
-
-    @staticmethod
-    def _expand_shape(shape_decl, writer):
-        if isinstance(shape_decl, str):
-            shape = tuple(writer.param[shape_decl])
-        elif hasattr(shape_decl, '__iter__'):
-            shape = tuple(writer.param[n] if isinstance(n, str) else n
-                          for n in shape_decl)
-        else:
-            shape = shape_decl
-
-        return shape
 
 
 class FileWriterMeta(type):
@@ -883,10 +315,6 @@ class FileWriterBase(object):
         self.trains.append(tid)
         self.timestamp.append(ts)
         self.flags.append(1)
-
-
-DS = Dataset
-trs_ = MultiTrainData
 
 
 class FileWriter(FileWriterBase, metaclass=FileWriterMeta):
