+++ tests/test_writer.py
@@ -1,15 +1,14 @@
-import h5py
 import os.path as osp
 import numpy as np
 from tempfile import TemporaryDirectory
 from testpath import assert_isfile
 
-from extra_data import RunDirectory, H5File
+from extra_data import RunDirectory
 from extra_writer import FileWriter, DS
 
 
 class MyFileWriter(FileWriter):
-    gv = DS('@ctrl', 'geom.fragmentVectors', (10,100), float)
+    gv = DS('@ctrl', 'geom.fragmentVectors', (10, 100), float)
     nb = DS('@ctrl', 'param.numberOfBins', (), np.uint64)
     rlim = DS('@ctrl', 'param.radiusRange', (2,), float)
 
@@ -20,7 +19,7 @@ class MyFileWriter(FileWriter):
     class Meta:
         max_train_per_file = 10
         break_into_sequence = True
-        #warn_on_missing_data = True
+        # warn_on_missing_data = True
         aliases = {
             'ctrl': '{detname}/x/y',
             'inst': '{detname}/x/y:output',
@@ -32,12 +31,11 @@ def test_writer():
     detname = 'MID_DET_AGIPD1M-1'
     ctrl_grp = f'{detname}/x/y'
     inst_grp = f'{detname}/x/y:output'
-    
 
     with TemporaryDirectory() as td:
         new_file = osp.join(td, 'test{seq:03d}.h5')
 
-        trains = list(range(100001,100026))
+        trains = list(range(100001, 100026))
         pulses = list(range(4, 16, 4))
         ntrain = len(trains)
         npulse = len(pulses)
@@ -61,7 +59,7 @@ def test_writer():
                 wr.add_train(tid, 0)
                 # add data
                 wr.add_train_data(gv=gv, nb=nbin, rlim=(0.003, 0.016))
-                
+
                 wr.tid = [tid] * npulse
                 wr.pid = pulses
                 wr.v = v
@@ -73,23 +71,24 @@ def test_writer():
             nseq = ntrain // tpf + int((ntrain % tpf) > 0)
         else:
             nseq, tpf = 1, ntrain
-            
+
         for seq in range(nseq):
             assert_isfile(new_file.format(seq=seq))
 
         with RunDirectory(td) as run:
-            np.testing.assert_array_equal(run.train_ids,
+            np.testing.assert_array_equal(
+                run.train_ids,
                 np.array(trains, dtype=np.uint64))
 
             assert ctrl_grp in run.control_sources
             assert inst_grp in run.instrument_sources
-            
+
             assert len(run.files) == nseq
             for i in range(nseq):
                 seq = int(run.files[i].filename[-6:-3])
                 t0, tN = seq * tpf, min(seq * tpf + tpf, ntrain)
                 np.testing.assert_array_equal(
-                    run.files[i].train_ids, 
+                    run.files[i].train_ids,
                     np.array(trains[t0:tN], dtype=np.uint64))
 
             nb = run.get_array(ctrl_grp, 'param.numberOfBins')
@@ -98,7 +97,7 @@ def test_writer():
 
             rlim = run.get_array(ctrl_grp, 'param.radiusRange')
             np.testing.assert_array_equal(rlim.values, np.broadcast_to(
-                np.array([0.003, 0.016], dtype=float), (ntrain,2)))
+                np.array([0.003, 0.016], dtype=float), (ntrain, 2)))
 
             frg = run.get_array(ctrl_grp, 'geom.fragmentVectors')
             np.testing.assert_array_equal(frg.values, np.broadcast_to(
